# openapi.yml
openapi: 3.1.0
info:
  title: Knjige i njihovi autori
  version: 3.0.0
paths:
...
  /api/books:
    get:
      tags:
        - books
      summary: Get books
      description: Get all books formatted
      operationId: getFromattedBooks
      responses:
        "200":
          description: Successful operation
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: No books found
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Not Found",
                      "message": "No books found",
                      "response": "",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "500":
          description: Internal Server Error while getting books
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Internal Server Error",
                      "message": "Error getting books",
                      "response": "",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
    post:
      tags:
        - books
      summary: Add book
      description: Add new book to the database
      operationId: addBook
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/Book/items"
      responses:
        "200":
          description: Successful operation
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "OK",
                      "message": "Book created",
                      "response": "GET /api/books/{id}",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "400":
          description: Invalid body or book
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Bad Request",
                      "message": "Invalid book",
                      "response": "",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "404":
          description: Author for book not found
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Not Found",
                      "message": "Author with key {key} not found",
                      "response": "GET /api/authors",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "409":
          description: Book with this key already exists
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ResponseWithBook"
                examples:
                  [
                    {
                      "status": "Conflict",
                      "message": "Book with this key already exists",
                      "response":
                        {
                          "number_of_pages": 127,
                          "title": 'Brodie''s Notes on Emily Bronte''s "Wuthering Heights" (Brodies Notes)',
                          "publish_date": "August 7, 1992",
                          "publishers": ["Palgrave Macmillan"],
                          "last_modified": "2022-11-15T20:13:03.789779",
                          "type": "edition",
                          "revision": 13,
                          "latest_revision": 13,
                          "authors":
                            [
                              {
                                "last_modified": "2008-08-19 18:21:52.259027",
                                "type": "author",
                                "name": "Norman T. Carrington",
                                "key": "OL1158982A",
                                "revision": 2,
                              },
                            ],
                          "key": "OL10551435M",
                          "subjects":
                            [
                              "English literature: literary criticism",
                              "Novels, other prose & writers: 19th century",
                              "Designed / suitable for A & AS Level",
                              "For National Curriculum Key Stage 4 & GCSE",
                              "Study guides, home study & revision notes",
                              "English",
                              "Children: Young Adult (Gr. 7-9)",
                            ],
                        },
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "500":
          description: Internal Server Error while adding book
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Internal Server Error",
                      "message": "Error putting book",
                      "response": "",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
    put:
      tags:
        - books
      summary: Add or update book
      description: Add new book to the database or update an existing book
      operationId: addOrUpdateBook
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/Book/items"
      responses:
        "200":
          description: Successful operation
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"

                examples:
                  [
                    {
                      "status": "OK",
                      "message": "Book updated/created",
                      "response": "GET /api/books/{id}",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "400":
          description: Invalid body or book
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Bad Request",
                      "message": "Invalid book",
                      "response": "",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "404":
          description: Author for book not found
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Not Found",
                      "message": "Author with key {key} not found",
                      "response": "GET /api/authors",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "500":
          description: Internal Server Error while adding or updating book
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Internal Server Error",
                      "message": "Error putting book",
                      "response": "",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
  /api/books/{id}:
    get:
      tags:
        - books
      summary: Get book
      description: Get book formatted
      operationId: getFromattedBook
      responses:
        "200":
          description: Successful operation
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Book/items"
        "404":
          description: Book with id {id} not found
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Not Found",
                      "message": "Book with id {id} not found",
                      "response": "GET /api/books",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "500":
          description: Internal Server Error while getting book
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Internal Server Error",
                      "message": "Error getting book data",
                      "response": "",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
    delete:
      tags:
        - books
      summary: Delete book
      description: Delete book from database
      operationId: deleteBook
      responses:
        "200":
          description: Successful operation
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ResponseWithBook"
                examples:
                  [
                    {
                      "status": "OK",
                      "message": "Book with id {id} deleted",
                      "response":
                        {
                          "publishers": ["IDG Books Worldwide"],
                          "number_of_pages": 331,
                          "latest_revision": 10,
                          "key": "OL824543M",
                          "authors": ["OL51749A"],
                          "title": "VCRs & camcorders for dummies",
                          "subjects":
                            [
                              "Home video systems.",
                              "Video tape recorders.",
                              "Camcorders.",
                              "Television -- Equipment and supplies.",
                            ],
                          "publish_date": "1995",
                          "last_modified": "2020-11-20T15:58:35.964496",
                          "type": "edition",
                          "revision": 10,
                        },
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "404":
          description: Book with id {id} not found
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Not Found",
                      "message": "Book with id {id} not found",
                      "response": "GET /api/books",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "500":
          description: Internal Server Error while deleting book
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Internal Server Error",
                      "message": "Error deleting book",
                      "response": "",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
  /api/authors:
    get:
      tags:
        - authors
      summary: Get authors
      description: Get all authors formatted
      operationId: getFromattedAuthors
      responses:
        "200":
          description: Successful operation
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Book/items/properties/authors"
        "404":
          description: No authors found
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Not Found",
                      "message": "No authors found",
                      "response": "",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
        "500":
          description: Internal Server Error while getting authors
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Response"
                examples:
                  [
                    {
                      "status": "Internal Server Error",
                      "message": "Error getting authors",
                      "response": "",
                      "timestamp": "2023-12-18T10:24:10.044Z",
                    },
                  ]
...
components:
  schemas:
    Book:
      "$schema": https://json-schema.org/draft/2019-09/schema
      "$id": https://github.com/fmarkulin/fer-or/blob/main/schema.json
      type: array
      default: []
      title: The Books and their authors Schema
      description: A list of books and their authors
      items:
        type: object
        title: A Books and their authors properties Schema
        required:
          - number_of_pages
          - title
          - publish_date
          - publishers
          - last_modified
          - type
          - revision
          - latest_revision
          - authors
          - key
          - subjects
        properties:
          number_of_pages:
            type: integer
            title: The number_of_pages Schema
            description: Number of pages in the book
            examples:
              - 127
              - 78
              - 192
              - 26
              - 200
              - 48
              - 298
              - 331
              - 96
          title:
            type: string
            title: The title Schema
            description: The title of the book
            examples:
              - Brodie's Notes on Emily Bronte's "Wuthering Heights" (Brodies Notes)
              - Aristophanes, ein Spiegel unserer Zeit.
              - War in Europe
              - VCRs & camcorders for dummies
          publish_date:
            type: string
            title: The publish_date Schema
            description: The date the book was published
            examples:
              - August 7, 1992
              - "1913"
          publishers:
            type: array
            title: The publishers Schema
            description: The publishers of the book
            items:
              type: string
              title: A publisher Schema
              examples:
                - Palgrave Macmillan
                - P. Schettlers
                - Presidio
            examples:
              - - Palgrave Macmillan
              - - P. Schettlers
              - - Presidio
              - - Palgrave Macmillan
                - Presidio
          last_modified:
            type: string
            title: The last_modified Schema
            description: The timestamp the book was last modified
            examples:
              - "2022-11-15T20:13:03.789779"
              - "2022-12-17T03:01:10.857196"
          type:
            type: string
            title: The type Schema
            description: The type of the book
            examples:
              - edition
          revision:
            type: integer
            title: The revision Schema
            description: The revision of the book
            examples:
              - 13
              - 3
              - 10
              - 5
              - 2
              - 6
          latest_revision:
            type: integer
            title: The latest_revision Schema
            description: The latest revision of the book
            examples:
              - 13
              - 3
              - 10
              - 5
              - 2
              - 6
          authors:
            type: array
            title: The authors Schema
            description: The authors of the book
            items:
              type: object
              title: A author Schema
              description: A author of the book
              required:
                - last_modified
                - type
                - name
                - key
                - revision
              properties:
                last_modified:
                  type: string
                  title: The last_modified Schema
                  examples:
                    - "2008-08-30 03:01:23.375753"
                    - "2008-04-29 15:03:11.581851"
                type:
                  type: string
                  title: The type Schema
                  description: The type of the entry
                  examples:
                    - author
                name:
                  type: string
                  title: The name Schema
                  description: The name of the author
                  examples:
                    - Norman T. Carrington
                    - A. B Kneisel
                    - Sidney C. Moody
                    - W. E. Hamilton
                    - Walker, John
                    - L. N. Carrier
                    - Rose, A.
                    - Gordon McComb
                    - Marie Jones
                    - Lorna Read
                key:
                  type: string
                  title: The key Schema
                  description: The key of the author
                  examples:
                    - OL1158982A
                    - OL4473583A
                    - OL581008A
                    - OL5295787A
                    - OL2555401A
                    - OL5288239A
                    - OL2153880A
                    - OL51749A
                    - OL500580A
                    - OL2802829A
                revision:
                  type: integer
                  title: The revision Schema
                  description: The revision of the author
                  examples:
                    - 2
                    - 1
              examples:
                - last_modified: "2008-08-19 18:21:52.259027"
                  type: author
                  name: Norman T. Carrington
                  key: OL1158982A
                  revision: 2
                - type: author
                  name: A. B Kneisel
                  last_modified: "2008-09-02 18:39:32.675717"
                  revision: 1
                  key: OL4473583A
            examples:
              - - last_modified: "2008-08-19 18:21:52.259027"
                  type: author
                  name: Norman T. Carrington
                  key: OL1158982A
                  revision: 2
              - - type: author
                  name: A. B Kneisel
                  last_modified: "2008-09-02 18:39:32.675717"
                  revision: 1
                  key: OL4473583A
          key:
            type: string
            title: The key Schema
            description: The key of the book
            examples:
              - OL10551435M
              - OL13994901M
          subjects:
            type: array
            title: The subjects Schema
            description: The subjects of the book
            items:
              type: string
              title: A subject Schema
              examples:
                - "English literature: literary criticism"
                - "Novels, other prose & writers: 19th century"
                - English
                - World War, 1939-1945 -- Campaigns -- Europe.
            examples:
              - - "English literature: literary criticism"
                - "Novels, other prose & writers: 19th century"
                - Designed / suitable for A & AS Level
                - For National Curriculum Key Stage 4 & GCSE
                - Study guides, home study & revision notes
                - English
                - "Children: Young Adult (Gr. 7-9)"
              - - Aristophanes
              - - World War, 1939-1945 -- Campaigns -- Europe.
                - World War, 1939-1945 -- Campaigns -- Europe -- Pictorial works.
        examples:
          - number_of_pages: 127
            title: Brodie's Notes on Emily Bronte's "Wuthering Heights" (Brodies Notes)
            publish_date: August 7, 1992
            publishers:
              - Palgrave Macmillan
            last_modified: "2022-11-15T20:13:03.789779"
            type: edition
            revision: 13
            latest_revision: 13
            authors:
              - last_modified: "2008-08-19 18:21:52.259027"
                type: author
                name: Norman T. Carrington
                key: OL1158982A
                revision: 2
            key: OL10551435M
            subjects:
              - "English literature: literary criticism"
              - "Novels, other prose & writers: 19th century"
              - Designed / suitable for A & AS Level
              - For National Curriculum Key Stage 4 & GCSE
              - Study guides, home study & revision notes
              - English
              - "Children: Young Adult (Gr. 7-9)"
          - revision: 3
            publishers:
              - P. Schettlers
            key: OL13994901M
            number_of_pages: 78
            type: edition
            latest_revision: 3
            title: Aristophanes, ein Spiegel unserer Zeit.
            subjects:
              - Aristophanes
            authors:
              - type: author
                name: A. B Kneisel
                last_modified: "2008-09-02 18:39:32.675717"
                revision: 1
                key: OL4473583A
            publish_date: "1913"
            last_modified: "2010-04-13T09:56:10.635564"
          - revision: 10
            type: edition
            key: OL1413776M
            publish_date: "1993"
            authors:
              - type: author
                name: Sidney C. Moody
                last_modified: "2008-08-30 03:01:23.375753"
                key: OL581008A
                revision: 2
            publishers:
              - Presidio
            title: War in Europe
            number_of_pages: 192
            last_modified: "2022-12-05T06:01:35.024861"
            subjects:
              - World War, 1939-1945 -- Campaigns -- Europe.
              - World War, 1939-1945 -- Campaigns -- Europe -- Pictorial works.
            latest_revision: 10
      examples:
        - - number_of_pages: 127
            title: Brodie's Notes on Emily Bronte's "Wuthering Heights" (Brodies Notes)
            publish_date: August 7, 1992
            publishers:
              - Palgrave Macmillan
            last_modified: "2022-11-15T20:13:03.789779"
            type: edition
            revision: 13
            latest_revision: 13
            authors:
              - last_modified: "2008-08-19 18:21:52.259027"
                type: author
                name: Norman T. Carrington
                key: OL1158982A
                revision: 2
            key: OL10551435M
            subjects:
              - "English literature: literary criticism"
              - "Novels, other prose & writers: 19th century"
              - Designed / suitable for A & AS Level
              - For National Curriculum Key Stage 4 & GCSE
              - Study guides, home study & revision notes
              - English
              - "Children: Young Adult (Gr. 7-9)"
          - revision: 3
            publishers:
              - P. Schettlers
            key: OL13994901M
            number_of_pages: 78
            type: edition
            latest_revision: 3
            title: Aristophanes, ein Spiegel unserer Zeit.
            subjects:
              - Aristophanes
            authors:
              - type: author
                name: A. B Kneisel
                last_modified: "2008-09-02 18:39:32.675717"
                revision: 1
                key: OL4473583A
            publish_date: "1913"
            last_modified: "2010-04-13T09:56:10.635564"
          - revision: 10
            type: edition
            key: OL1413776M
            publish_date: "1993"
            authors:
              - type: author
                name: Sidney C. Moody
                last_modified: "2008-08-30 03:01:23.375753"
                key: OL581008A
                revision: 2
            publishers:
              - Presidio
            title: War in Europe
            number_of_pages: 192
            last_modified: "2022-12-05T06:01:35.024861"
            subjects:
              - World War, 1939-1945 -- Campaigns -- Europe.
              - World War, 1939-1945 -- Campaigns -- Europe -- Pictorial works.
            latest_revision: 10
...
    Response:
      type: object
      properties:
        status:
          type: string
          examples: [OK, Bad Request, Not Found, Internal Server Error]
        message:
          type: string
          examples:
            - Book with id {id} deleted
            - Invalid book
            - Author with key {key} not found
            - Error deleting book
            - Error getting book data
            - Error getting authors
            - Error putting book
        response:
          type: string
          examples:
            - GET /api/books
            - GET /api/authors
            - GET /api/books/{id}
            - ""
        timestamp:
          type: string
          examples:
            - "2023-12-18T10:24:10.044Z"
            - "2023-12-18T10:24:10.044Z"
            - "2023-12-18T10:24:10.044Z"
            - "2023-12-18T10:24:10.044Z"
            - "2023-12-18T10:24:10.044Z"
            - "2023-12-18T10:24:10.044Z"
            - "2023-12-18T10:24:10.044Z"
            - "2023-12-18T10:24:10.044Z"
            - "2023-12-18T10:24:10.044Z"
            - "2023-12-18T10:24:10.044Z"
...